added invocation tests
- uses https://github.com/Valodim/zsh-capture-completion for zsh
- extracted script fixtues
