+++ example/_test/invoke_bash
@@ -1,0 +1,10 @@
+#!/bin/bash -i
+#set -x
+source <(example _carapace bash)
+COMP_LINE="$1"
+COMP_POINT=${#1}
+COMP_WORDS=($COMP_LINE'')
+COMP_CWORD=$((${#COMP_WORDS[@]}-1))
+
+_example_completions
+( IFS=$'\n'; echo "${COMPREPLY[*]}" )
