+++ example/_test/invoke_powershell
@@ -1,0 +1,4 @@
+#!/usr/bin/pwsh
+
+example _carapace powershell | out-string | invoke-expression
+[System.Management.Automation.CommandCompletion]::CompleteInput("$($args[0])", $args[0].length, $null).CompletionMatches
